requirements.txt
streamlit>=1.36
tensorflow-cpu==2.20.0
numpy>=1.26
pillow
matplotlib
